A first attempt at GC yield prediction.
This attempts to predict the yield of a major GC, in terms of number of
reclaimed objects. A GC is presumed to be maximally effective when there
are relatively few live objects, which often occurs between phases of a
program. This diff allows garbage collection to be delayed until the
estimated ratio of live to dead objects falls below a certain threshold:
a "plague" has occurred. This should increase heap size while reducing
the number of major collections and their duration.