[sgen] Region allocator.
The initial idea for this was to implement optimistic stack allocation:
to allocate an object on the stack, and only move it to the heap when a
reference to it would escape its scope. The idea behind this is that not
only do most objects die young, but most objects die in the method where
they were allocated.

However, this presented some challenges. We cannot actually move objects
from the stack to the heap, because it would require a read barrier to
fix references to escaped objects.

In light of that, this code implements a "region allocator" inside the
nursery. When entering a method, we add the current TLAB pointer to a
"region stack". If no reference to an object in the current region has
escaped the current region, then when we exit the method by returning or
throwing an exception, we can clear the memory and reset the TLAB
pointer. Then we can reuse that portion of the TLAB.

If a reference to an object in the current region has escaped, then we
"stick" the current region, and discard all information about higher
region stack frames. We use the allocator and write barrier to detect
escaped references. A reference escapes if any of the following is true:

 * Its type has a finalizer

 * It is reachable from an object returned or thrown from a method
   (FIXME: not fully implemented)

 * It is written into:

    * The major heap

    * The nursery outside the current TLAB

    * A higher region

    * A higher stack frame

This also generates a write barrier for static field initializers.